feat(widgets): improve format and fonts of Item widgets empty view
