+++ PocketKit/Sources/ItemWidgetsKit/Common/Views/Configuration/Style+ItemWidgets.swift
@@ -7,4 +7,5 @@ import Textile
 extension Style {
     static let domain: Style = .header.sansSerif.p6.with(color: .ui.grey8)
     static let widgetHeader: Style = .header.sansSerif.h8.with(color: .ui.coral2)
+    static let emptyWidgetMessage: Style = .header.sansSerif.p2.with(color: .ui.grey8).with(alignment: .center)
 }
